mirros multi-turn sample more closely
