--- samples/javascript_nodejs/09.adaptive-card-prompt/dialogs/mainDialog.js
@@ -1,50 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-
-const { ComponentDialog, DialogSet, DialogTurnStatus, WaterfallDialog } = require('botbuilder-dialogs');
-const { AdaptiveCardDialog, ADAPTIVE_CARD_DIALOG } = require('./adaptiveCardDialog');
-
-const MAIN_DIALOG = 'MAIN_DIALOG';
-const WATERFALL_DIALOG = 'WATERFALL_DIALOG';
-
-class MainDialog extends ComponentDialog {
-    constructor() {
-        super(MAIN_DIALOG);
-
-        this.addDialog(new AdaptiveCardDialog());
-        this.addDialog(new WaterfallDialog(WATERFALL_DIALOG, [
-            this.initialStep.bind(this),
-            this.finalStep.bind(this)
-        ]));
-
-        this.initialDialogId = WATERFALL_DIALOG;
-    }
-
-    /**
-     * The run method handles the incoming activity (in the form of a TurnContext) and passes it through the dialog system.
-     * If no dialog is active, it will start the default dialog.
-     * @param {*} turnContext
-     * @param {*} accessor
-     */
-    async run(turnContext, accessor) {
-        const dialogSet = new DialogSet(accessor);
-        dialogSet.add(this);
-
-        const dialogContext = await dialogSet.createContext(turnContext);
-        const results = await dialogContext.continueDialog();
-        if (results.status === DialogTurnStatus.empty) {
-            await dialogContext.beginDialog(this.id);
-        }
-    }
-
-    async initialStep(stepContext) {
-        return await stepContext.beginDialog(ADAPTIVE_CARD_DIALOG);
-    }
-
-    async finalStep(stepContext) {
-        return await stepContext.endDialog();
-    }
-}
-
-module.exports.MainDialog = MainDialog;
-module.exports.MAIN_DIALOG = MAIN_DIALOG;
