+++ samples/javascript_nodejs/09.adaptive-card-prompt/index.js
@@ -9,17 +9,17 @@ const restify = require('restify');
 const { BotFrameworkAdapter, MemoryStorage, UserState, ConversationState } = require('botbuilder');
 
 const { DialogAndWelcomeBot } = require('./bots/dialogAndWelcomeBot');
-const { MainDialog } = require('./dialogs/mainDialog');
+const { AdaptiveCardPromptDialog } = require('./dialogs/adaptiveCardPromptDialog');
 
 // Read environment variables from .env file
 const ENV_FILE = path.join(__dirname, '.env');
 require('dotenv').config({ path: ENV_FILE });
 
 // Create HTTP server
-let server = restify.createServer();
+const server = restify.createServer();
 server.listen(process.env.port || process.env.PORT || 3978, function() {
     console.log(`\n${ server.name } listening to ${ server.url }`);
-    console.log(`\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator`);
+    console.log('\nGet Bot Framework Emulator: https://aka.ms/botframework-emulator');
 });
 
 // Create adapter.
@@ -38,7 +38,7 @@ const userState = new UserState(memoryStorage);
 const conversationState = new ConversationState(memoryStorage);
 
 // Create the main dialog.
-const dialog = new MainDialog();
+const dialog = new AdaptiveCardPromptDialog();
 const bot = new DialogAndWelcomeBot(conversationState, userState, dialog);
 
 // Catch-all for errors.
@@ -46,7 +46,7 @@ adapter.onTurnError = async (context, error) => {
     // This check writes out errors to console log .vs. app insights.
     console.error(`\n [onTurnError]: ${ error }`);
     // Send a message to the user
-    await context.sendActivity(`Oops. Something went wrong!`);
+    await context.sendActivity('Oops. Something went wrong!');
     // Clear out state
     await conversationState.load(context);
     await conversationState.clear(context);
